Adjusted Card Class to single Shape + docs
Performance was okay on apple, but android was really laggy - could not
even use the app.  So freaked out.  Then commented out the cards and
performance was great!  So redesigned cards to work with one Shape
rather than 169 rectangles.  Created a ZIM hitTestGrid function to
control the shape and the performance is great!