Now adjusted the link to danzen
Meant to do it before but got mixed up.  Have now officially moved to
editing in the git folder.  Initially, this makes me claustrophobic.  So
at the moment, I just build for the web and when close to getting to
mobile post to git - but perhaps I will change my mind - and of course,
I am working alone.